DB: Use ID as offset in item queries
