+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is almost based on [Keep a Changelog](https://keepachangelog.com/en/1
 - Item list throwing error for folder and "all items"
 - Articles with high IDs can be placed lower than articles with low IDs (#1147)
 - Feeds are accidentally moved on rename
+- Item list not using ID for offset
 
 ## [15.3.2] - 2021-02-10
 No changes compared to RC2
